+++ src/ariadne_findcache.c
@@ -31,23 +31,23 @@
 extern "C" {
 #endif
 /******************************************************************************/
-static long find_cache[4][4];
+static int32_t find_cache[4][4];
 
 /******************************************************************************/
-long triangle_brute_find8_near(long pos_x, long pos_y)
+int32_t triangle_brute_find8_near(int32_t pos_x, int32_t pos_y)
 {
-    long cx = pos_x >> 14;
+    int32_t cx = pos_x >> 14;
     if (cx < 0)
         cx = 0;
     if (cx > 3)
         cx = 3;
-    long cy = pos_y >> 14;
+    int32_t cy = pos_y >> 14;
     if (cy < 0)
         cy = 0;
     if (cy > 3)
         cy = 3;
    // Try sibling
-    long tri_id;
+    int32_t tri_id;
     if (cx-1 >= 0)
     {
         tri_id = find_cache[cy][cx-1];
@@ -87,20 +87,20 @@ long triangle_brute_find8_near(long pos_x, long pos_y)
     return tri_id;
 }
 
-static long triangle_find_cache_get(long pos_x, long pos_y)
+static int32_t triangle_find_cache_get(int32_t pos_x, int32_t pos_y)
 {
-    long cache_x = (pos_x >> 14);
+    int32_t cache_x = (pos_x >> 14);
     if (cache_x > 3)
         cache_x = 3;
     if (cache_x < 0)
         cache_x = 0;
-    long cache_y = (pos_y >> 14);
+    int32_t cache_y = (pos_y >> 14);
     if (cache_y > 3)
         cache_y = 3;
     if (cache_y < 0)
         cache_y = 0;
 
-    long ntri = find_cache[cache_y][cache_x];
+    int32_t ntri = find_cache[cache_y][cache_x];
     if (get_triangle_tree_alt(ntri) == NAV_COL_UNSET)
     {
         ntri = triangle_brute_find8_near(pos_x, pos_y);
@@ -115,14 +115,14 @@ static long triangle_find_cache_get(long pos_x, long pos_y)
 
 }
 
-static void triangle_find_cache_put(long pos_x, long pos_y, long ntri)
+static void triangle_find_cache_put(int32_t pos_x, int32_t pos_y, int32_t ntri)
 {
-    long cache_x = (pos_x >> 14);
+    int32_t cache_x = (pos_x >> 14);
     if (cache_x > 3)
         cache_x = 3;
     if (cache_x < 0)
         cache_x = 0;
-    long cache_y = (pos_y >> 14);
+    int32_t cache_y = (pos_y >> 14);
     if (cache_y > 3)
         cache_y = 3;
     if (cache_y < 0)
@@ -130,9 +130,9 @@ static void triangle_find_cache_put(long pos_x, long pos_y, long ntri)
     find_cache[cache_y][cache_x] = ntri;
 }
 
-void triangulation_init_cache(long tri_idx)
+void triangulation_init_cache(int32_t tri_idx)
 {
-    for (long i = 0; i < 4; i++)
+    for (int32_t i = 0; i < 4; i++)
     {
         find_cache[i][0] = tri_idx;
         find_cache[i][1] = tri_idx;
@@ -141,19 +141,19 @@ void triangulation_init_cache(long tri_idx)
     }
 }
 
-long triangle_find8(long pt_x, long pt_y)
+int32_t triangle_find8(int32_t pt_x, int32_t pt_y)
 {
     NAVIDBG(19,"Starting");
     //TODO PATHFINDING triangulate_area sub-sub-sub-function
     int32_t ntri = triangle_find_cache_get(pt_x, pt_y);
-    for (unsigned long k = 0; k < TRIANLGLES_COUNT; k++)
+    for (uint32_t k = 0; k < TRIANLGLES_COUNT; k++)
     {
         int eqA = triangle_divide_areas_s8differ(ntri, 0, 1, pt_x, pt_y) > 0;
         int eqB = triangle_divide_areas_s8differ(ntri, 1, 2, pt_x, pt_y) > 0;
         int eqC = triangle_divide_areas_s8differ(ntri, 2, 0, pt_x, pt_y) > 0;
 
         int32_t ncor = 0;
-        long nxcor = 0; // Used only to verify if pointed_at8() didn't failed
+        int32_t nxcor = 0; // Used only to verify if pointed_at8() didn't failed
         switch ((eqC << 2) + (eqB << 1) + eqA)
         {
         case 1:
@@ -202,14 +202,14 @@ long triangle_find8(long pt_x, long pt_y)
  * @param output_corner_index
  * @return
  */
-TbBool point_find(long pt_x, long pt_y, int32_t *output_triangle_index, int32_t *output_corner_index)
+TbBool point_find(int32_t pt_x, int32_t pt_y, int32_t *output_triangle_index, int32_t *output_corner_index)
 {
-    long tri_idx = triangle_find8(pt_x << 8, pt_y << 8);
+    int32_t tri_idx = triangle_find8(pt_x << 8, pt_y << 8);
     if (tri_idx < 0)
     {
         return false;
     }
-    for (long cor_id = 0; cor_id < 3; cor_id++)
+    for (int32_t cor_id = 0; cor_id < 3; cor_id++)
     {
         struct Point* pt = get_triangle_point(tri_idx, cor_id);
         if ((pt->x == pt_x) && (pt->y == pt_y))
