+++ src/ariadne_wallhug.h
@@ -39,8 +39,8 @@ struct Navigation {
   unsigned char wallhug_retry_counter;
   unsigned char wallhug_state;
   unsigned char push_counter;
-  long dist_to_final_pos;
-  long distance_to_next_pos;
+  int32_t dist_to_final_pos;
+  int32_t distance_to_next_pos;
   int32_t angle;
   SubtlCodedCoords first_colliding_block;
   SubtlCodedCoords second_colliding_block;
@@ -60,8 +60,8 @@ enum WallHugSideState {
 /******************************************************************************/
 
 /******************************************************************************/
-long slab_wall_hug_route(struct Thing *thing, struct Coord3d *pos, long max_val);
-long get_next_position_and_angle_required_to_tunnel_creature_to(struct Thing *creatng, struct Coord3d *pos, PlayerBitFlags crt_owner_flags);
+int32_t slab_wall_hug_route(struct Thing *thing, struct Coord3d *pos, int32_t max_val);
+int32_t get_next_position_and_angle_required_to_tunnel_creature_to(struct Thing *creatng, struct Coord3d *pos, PlayerBitFlags crt_owner_flags);
 SubtlCodedCoords dig_to_position(PlayerNumber plyr_idx, MapSubtlCoord basestl_x, MapSubtlCoord basestl_y, SmallAroundIndex direction_around, TbBool revside);
 TbBool slab_good_for_computer_dig_path(const struct SlabMap *slb);
 short get_hug_side_options(MapSubtlCoord stl1_x, MapSubtlCoord stl1_y, MapSubtlCoord stl2_x, MapSubtlCoord stl2_y, SmallAroundIndex direction, PlayerNumber plyr_idx,
